+++ CHANGELOG.md
@@ -13,6 +13,9 @@ v5.4.0
 - Set initial condition check points files to zero (0) (!1535)
 - Prevent file overwrite with restart for the Compressible flow solver (!1540)
 
+**IncNavierStokesSolver**
+- Register SolverType in SessionReader (!1541)
+
 **Miscellaneous**
 - Tidy-up Parareal initial condition output (!1534)
 - Remove double entry in documentation and slightly tidy-up code (!1530)